requirements.txt
Django==5.2.1
djangorestframework==3.15.2
celery==5.4.0
django-celery-beat==2.7.0
python-dotenv==1.0.1
